Move HTTP features from router-http to linkerd-protocol-http.
The router-http package changes tracing configuration and reports errors using
`Buoy-` headers.  This is undesirable, as router-http should be as close to
Finagle as possible, and should not impose changes like this.

We change this to use the `l5d-` header prefix, and move many of the HTTP
features into the linkerd-protocol-http project.  Changes include:
- http-specific trace initialization
- access logging
- error code responses
- http status code measurement