Simplify router stats scoping under rt/$label
- all router stats are scoped under rt/$label
- client stats are scoped under rt/$label/dst/id/$id
- server stats are scoped under rt/$label/srv/$host/$port
- later, additional client stats will be scoped under rt/$label/dst/path/$path.

The admin interface has been changed so that metrics processing results in
information about routers (rather than just clients and servers).  This data is
used to power the same UI without changes, although UI changes should follow
this in a followup review.

If a Router is configured with a StatsReceiver or Tracer, this configuration is
not currently propagated to Servers.  This Router configuration should be
shared with its servers so that defaults may be specified and so that, later,
these may be configured in a per-router fashion.  This change allows Tracers
and scoped StatsReceivers to be shared from a Router to its clients and
servers.

Fixes #14