API Endpoint for configured routers
Admin pages need a list of routers in order to be able to switch between them. I was originally going to just include this info as part of the html payload, but found it awkward to have to pass the linker all the way to the adminHtml rendering method. It currently renders only a small amount of router information, but we can either 1) add more to the serializer as needed or 2) replace with a config serializer once we decide on our config parsing method