Introduce the PathMatcher utility
PathMatcher is intended to provide a way to process paths from a configuration
file without exposing a richer programming language.

PathMatcher is initialized with an expression string that can be used to
process paths in terms of their positional elements. A positional parameter may
be in the forms:
- $1
- ${1}
- $01
- etc, no math though

Any of these would refer to "foo" in the path "/foo/bar"; and an expression
like "$2.$1.com", this would expand to "bar.foo.com".