TracerConfig and TracerInitializer
I want to use my own tracer for linkerd routers and configure it
similar to how namers are configured.

My proposed solution is to add a top level `tracers` section to
the config. Each tracer requires a kind to be specified. See
TestTracer for an example on how I plan to use it