Establish a new client for each Kubernetes namespace.
Currently, the k8s namer builds a single client at configuration-time.  In
theory, this is fine, but we have observed odd hanging behavior, as described
by #83.

Anecdotally, establishing a new client for each namespace has resolved this
issue in the past.  This does not provide a real _fix_ for the issue, though it
may offer some relief for the issue while we establish a more robust fix.