Allow NamerInitializer modules to provide NameInterpreters.
In order to outsource name resolution (i.e. to another service), we need to be
able to provide a configured NameInterpeter implementation.  Rather than
complicate the configuration file to explicitly differentiate NameInterpreter
and Namer configurations, they are merged into the `namers` field.

If multiple NameInterpeter configurations are specified or if Namers are
specified along with a NameInterpreter, an error is thrown.