Add module to display summary of router stats.
The router modules will be controlled by a router controller which will handle getting data and distributing it to the modules which will do further processing on the data pertinent to them, and then render the result.

Also add a skeletal implementation of RouterServers to demonstrate how I think RouterController could be used.