Query util for clientside metrics gathering
While routers.js does a great job of creating a hierarchical view
of routers, servers, and clients, we're finding an increasing need
to make calls for metrics across these boundaries, which results in
a lot of mapping/filtering that is not always very readable.

This change introduces a module that builds a regex for querying
across a list of metrics.