linkerd-admin: Calculate success rate purely over success + failure
`requests` is measured in a different place from `success` and failure`.
`requests` counts the number of requests sent, not completed.  Because of this,
`success` can easily be greater than `requests`, which leads to odd success
rate measurements.

This is easily fixed by calculating total requests as `success + failure`.

default success and failures values to 0