Introduce DtabStore.Proxy type to support store-agnostic validation (#252)
Introduce DtabStore.Proxy type to support store-agnostic validation

We'll need to be perform write validation in an interface- and store-agnostic
way.

In order to support this, a DtabStore.Proxy type can be used to wrap DtabStores
with behavior. A special DtabStore.Validator type is introduced to provide a
basis for write validation.