Simplify address observation (#305)
Simplify address observation
Since address observations are a flatMap of bind with the Var[Addr], they
should always be correct, even if the NameTree becomes Neg and then comes back
with a new Var[Addr] (as in the case where a concrete name is deleted and then
recreated).  This means we can simplify the code by not evicting cached
observations when the NameTree becomes Neg.