Fix etcd event observation
Before, event observation used the maximum modifiedIndex from the returned
set of nodes. There are no guarantees that this index is anywhere near current,
and so the watch could replay the history of updates or fail with an unknown
index.

By using etcd index (as returned in the header), we can ensure we're watching
from a recent index.