Prevent re-use of sessions across retries (#413)
* Introduce a FactoryToService module into the path stack

The retry filter may issue multiple requests on the underlying service factory.
This interaction needs to be managed by a FactoryToService module so that the
underlying service (connection) is released to finagle between requests so
that, for instance, it may be closed in the case of HTTP/1.0 requests.

Fixes #409

* Move acquisition failure tracing below retries.

Now that factory acquisition is pushed below retries, the client factory lookup
always succeeds.  In order to record acquisition failures, recording must be
pushed below retries.

* closes aren't synchronous, so we can't rely on them to be timely in CI.

* Fixup commentary

* Move NoBrokersAvailable lifting down under the path stack

With the prior change, NoBrokersAvailableExceptions would not be decorated with
Dtab information because they are thrown at request time and not service
acquisition time (due to factoryToService). This change moves NoBrokers
decoration below the path stack.

* improve DstBindingFactory testing in face of factory-to-service