Use a static RoutingService in RoutingFactory (#423)
RoutingFactory currently builds a new RoutingService for each connection. The
server-side connection is passed into the DstBindingFactory.  This is wasteful,
considering that none of the server-side connection state matters to the
router. By simplifying this to use a static RoutingService with a nil
connection, we see substantial performance improvements in the face of
serverside connection churn.