linkerd-protocol-http: Handle client exceptions in the path stack
Problem

When an exception is raised in the linkerd client, it is thrown into the
server.  This means that client connection errors are measuerd as server
connection errors in stats, etc.

Solution

Handle exceptions at the top of the router's path stack so that the server
isn't exposed to client-generated exceptions.