Delay service closure while streams are open
With the introduction of retries, a subtle bug was introduced to HTTP session
management. When a server responded with `Transfer-encoding: chunked` and
`Connection: close`, connections could be incorrectly returned to the
connection pool.

By applying Finagle's DelayedReleaseService after FactoryToService, connections
are not relinquished to the connection pools until the response stream
completes.

Fixes #482.
More details at twitter/finagle#517.