Configure server-side timeouts
Router timeouts are only applied in the client stack, which is below retries,
etc. This means that retries could persist forever, since they would never
exceed a global timeout.

To fix this, the server is now configured with the router's timeout value, to
impose a global timeout requirement.

In the future, it may make sense to admit client and server-specific timeouts.