Support a Netty4 engine for HTTP routers
Finagle-6.36 introduces a usable netty4 backend for HTTP. linkerd users should
be able to opt-in to this new backend so that it may be introduced in a
controlled manner (until it's proven to be a suitable default).

Client and Server sections of the HTTP router configuration admit an `engine`
block that controls which HTTP implementation is used.