Handle "too old" error when re-establishing Kubernetes watches (#473)
* Handle "too old" error when re-establishing Kubernetes watches

When a watch against the k8s API is terminated (because of a time-out, connection failure, etc), it must be re-established. Kubernetes only allows watching from a "recent" version, so when re-establishing the watch, it's possible it will be rejected with a "resource too old" error. In this case, it's necessary to refresh the resource list to verify no updates were missed, and then re-establish the watch from the newly-received version number.

* Add tests for k8s "Resource too old" error