An optional recursive mode for ConsulDtabStore
This change gives an option to enable "composite" dtabs with Consul.
Given the namespace ConsulDtabStore will fetch all the keys
below the namespace recursively and combine them in lexicographical
order of keys while dropping empty and broken ones.
It watches for changes the whole tree of keys.
This mode is optional and off by default.