# This is a combination of 3 commits. # The first commit's message is: Add maxResponseKB to http client configuration
By default, Finagle and Netty set a 10MB limit for an unchunked response. This
is done to protect against having to buffer arbitrarily large payloads, which
can lead to performance degradation.

This change introduces a new configuration option for HTTP clients, allowing
this limit to be increased on a per-router basis.

# The 2nd commit message will be skipped:

#	wip

# This is the 3rd commit message:

Add configuration options for most of the HTTP codec's tunable parameters