Add configuration options for most of the HTTP codec's tunable parameters
By default, Finagle and Netty set a 5MB limit for unchunked bodies. This is
done to protect against having to buffer arbitrarily large payloads, which can
lead to performance degradation.

This change introduces a new configuration option for HTTP clients, allowing
the memory constraints of the HTP codec to be controlled on a per-router basis.