modify the delegate api to provide bound addresses
problem: the http /delegate endpoint would not provide bound addresses
unless a namespace parameter was provided.

solution: modify the behavior of /delegate to use available namers when
a namespace is not provided.

note: this does not change the behavior of the /delegate/[namespace]/
api, though it may be prudent in the future to harmonize or converge
these code paths.