Fix spurious test logging
MarathonTest emits many log messages indicating that "marathon.mesos" cannot be
resolved in DNS.  Use "localhost" in test marathon configs.

ThriftNamerEndToEndTest continually logs errors for the life of the process,
making it miserable to run `test` in an sbt console.  To fix this, we ensure
that the observation is released before the test completes.