Chunked requests must not be considered retryable (#601)
* Chunked requests must not be considered retryable

When an HTTP router is configured with the Idempotent repsonse classifier, it
may consider a request with a body (e.g. PUTs) as retryable.  However, chunked
bodies may not be replayed, since the stream is not preserved by linkerd.

To fix this, we only treat a request as retryable when it does not have a
chunked body.

Fixes #599

* +docs

* Make the HttpConfig enforce chunking retry policy