Fix k8s ApiTest
In #619, we made our retry behavior more liberal. However, some of the ApiTests
were not changed to reflect this increased retry behavior.  Furthermore, now
that we swallow and log exceptions in the retry filter, we need an explicit
mechanism to signal test failure to the test harness.

To fix this:
1. stop serving mocked responses to prevent infinite looping
2. we capture and re-raise exceptions thrown (i.e. by assert statements) manually
3. manually close streams in all tests

Fixes #633