Lazy, managable consul resolution (#632)
* Move SvcCache and DcCache into their own files.

The consul namer is a little difficult to reason about as one big intermingled
class. We begin to simplify this by moving DcCache and SvcCache into their own
package-local classes.

* Stop issuing redundant local agent requests.

Previously, the domain was resolved from consul on every service addition.
Instead, this should be modeled as a single activity that is

Furthermore, on DcCache updates, the dc cache was updated for each individual
service addition and removal.  Instead, we batch the DC update to be only once
per API response.

Furthermore, access & update of `index` variables was not thread safe for
visibility. `index` should ve volatile.

Finally, some minor allocation improvements were made:
- val'ing some anonymous functions, where convenient
- Only build address metadata for bound addresses

* Split ConsulNamer into Tagged, Untagged, and Base.

The ConsulNamer doesn't need to determine whether a tag should be included on
each and every lookup. Instead, we can determine this once at
configuration-time and construct a Namer that is statically configured to
produce service keys.

ConsulNamer.Base has all of the common logic shared between namer
implementations.

* ConsulNamer.Base => LookupCache

* Make it so that addr caching can be torn down.

It seems that the SvcCache leaks observations. After a service is removed from
a DC, there was no explicit mechanism to stop the service's request loop in the
SvcCache.  Now, cancellation is available on both DcCache and SvcCache.

* add an explicit mutex object in DcCache; cleanup looping

* Make SvcCache a simple Var.async[Addr.

Preiviously, the consul namer would eagerly start observing all services,
always. This works fine in small clusters but maybe be problematic in large
clusters.

We can fix this by creating a Var.async from the underlying service polling
loop.  Now, the namer will only query consul for services that linkerd is
trying to route on.

* SvcAddr: remove stateful `index` variable.

* fix comment

* Refactor DcCache to DcServices.

DcServices simply exposes an `Activity[Map[SvcKey, Var[Addr]]`, which is built
using an async Var that is initiated lazily and may be torn down when not
observed.

* Review feedback

- extract metadata construction into a helper.  Reuse metadata across addr updates.
- consistent error handling in DcServices and SvcAddr.
- commentary
- fix ConsulNamerTest to properly simulate missing DCs

* Fix metadata reuse; propagate no index failures.