Convert namerd docs to use tables.
Problem: namerd docs are not as well formatted as linkerd docs.
Solution: format them.

For future work I'd really like to see more examples in this documentation, as well
as documentation of the interface apis themselves.

I also have a pie-in-the-sky dream that all the configs standardize on
milliseconds, and be suffixed with "Ms". :)