k8s transformers (#669)
k8s transformers

Adds support for linkerd-to-linkerd deployments in k8s when linkerd is deployed
as a daemonset. This is accomplished by adding two NameTree transformers:
io.l5d.k8s.daemonset and io.l5d.k8s.localnode. For linkerd-to-linkerd
deployments, one should add the localnode transformer to the incoming router and
the daemonset transformer (configured with the linkerd daemonset) to the
outgoing router.