Gracefully detect incoming prior knowledge requests
Plaintext h2 servers may accept either HTTP/1 upgrade messages or prior
knowledge h2 requests.  ServerUpgradeHandler detects which type of message is
being received by looking at the first bytes of the message to determine if
includes the magic prior knowledge connection preface. If the message is not
prefaced appropriately, it is treated as an HTTP/1.1 request.

Rename param.PriorKnowledge to param.ClientPriorKnowledge, since this parameter
is only taken into account when a client initiates a connection.