Refactor h2 server to allow proper flow control
In the initial implementation of h2, our managed multiplexing in the Netty
stream so that H2 listeners created a Transport for each client stream (and not
per-connection, as is done for other protocols).

We also learned that Netty's (experimental) Http2MultiplexCodec has
flow-control behavior that do not currently play well with linkerd's model.
Specifically:
- Write promises may be satisfied before the entire data frame is written.
- WINDOW_UPDATE messages are automatically issued by the codec as soon as a
  data frame is read. In linkerd, we want to defer this update until the data
  frame has been completely forwarded.

The combination of these issues makes it difficult if not impossible for
linkerd to manage flow control properly.

This change moves linkerd's H2 server from netty's Http2Codec to
Http2FrameCodec.  The FrameCodec provides only the encoding/decoding subset of
features, and is already used by our client.  The Netty4ClientStreamTransport
is generalized as the Netty4StreamTransport, which is now used by both clients
and servers.

Now, h2 listeners create a Netty4StreamTransport for each client connection.
Netty4ServerDispatcher implements multiplexing logic---similar, but inverse, to
that of Netty4ClientDispatcher.  This allows linkerd to manage flow control
explicitly and makes the Client and Server code paths more congruent.

The h2 end-to-end test has been extended to validate flow control behavior.

Fixes #730.