Admit HTTP/1 upgrade requests in h2 plaintext servers (#787)
In 59df8f2, ServerUpgradeHandler was disabled to accomodate the new
multiplexing regime, requiring that all plaintext clients send a prior
knowledge request.

This change introduces an adapter, io.netty4...Http2FrameCodecServerUpgrader,
that instantiates an Http2ServerUpgradeCodec with a Http2FrameCodec so that the
server properly accepts HTTP/1 upgrade requests in addition to prior-knowledge
requests.

Fixes #765