Move common rewriting namers from router packages to namer-core (#802)
Several of the io.bouyant.http rewriting namers are inaccessible in namerd
because they are in router-http. Furthermore, the io.buoyant.rinet namer was
packaged in router-core.

The rewriting namers can simply be moved into namer-core (even though "http" is
in the class package name) because there are no actual HTTP-specific
dependencies.

Now these namers may be referenced when using namerd.

Fixes #789