Introduce the io.linkerd.mesh gRPC interface
Problem

namerd's thriftmux interface, while multiplexed, suffers from lack of
streaming primitives in thriftmux. The HTTP interface, on the other
hand, may stream messages but is not multiplexed and is JSON-encoded.

Solution

A new gRPC interface, io.linkerd.mesh, is introduced to eventualy
supersede the existing namerd read interfaces. Furthermore, this
interface will be expanded to cover a variety of additional linkerd
management concerns.

A good deal of testing and validation remains to be done on this
implementation. It is only suitable for evaluation purposes at this
time.