[h2] Refuse unroutable streams (#835)
When linkerd encounters an error---because a request is unroutable, or because
the downstream client fails (e.g. with a Reset as the response is
received)---linkerd currently lifts this failure into the application to an
HTTP/2 response.

Instead, we should use stream-level error codes to signal that linkerd was
unable to open a remote connection.