Add "originator" router parameter
Adds a new parameter for configuring routers: originator. If the
originator parameter is set to true, it indicates that this router is
the first hop in a linker-to-linker request. Routers that are configured
as originators will emit an additional stat (rt/label/originator == 1.0)
that can be used to deduplicate stats that are produced by 2 different
routers in the process of a serving 1 linker-to-linker request.