Limit the namerd mux client to a single endpoint
The thrift protocol implemented by namerd is designed for subsequent requests
to be sent to different nodes, as there are no guarantees that the 'stamp'
values used are shared across nodes. This watching behavior should later be
moved to a streaming interface. In the meantime, we use the Aperture load
balancer with an unlimited load so that only a single instance will be
considered. When a namerd instance is unavailable, another address in the pool
may be considered.

Fixes #295