Support the RFC 7329 `Forwarded` header
Problem

linkerd includes a `ForwardedFilter` but (1) it is disabled and the
configuration exposes no way to enable it, and (2) it is not quite
RFC-compliant.

Solution

Support the `addForwardedHeader` configuration object on HTTP servers. By
default, no `Forwarded` header is added to requests. When enabled, a random
string is used for each of the `by` and `for` endpoints. Users may configure
`by` and `for` to be labeled as: an IP, an IP:PORT, a random string, the parent
router's label, or a static label.

Fixes #902