Introduce io.buoyant.grpc.runtime.EventStream
Problem

It's common to model a Var/Activity over an RPC stream. It's awkward to do this
with the AsyncQueue-backed Stream/Provider because the future-oriented
back-pressure model does not interact well with the Event.respond API.

Solution

Provide a new Stream type, EventStream, that maintains only the most
recently observed Event value. Stream receivers may miss intermediate states,
but the stream should generally behave better when the consumer is slow.