Change the internal Delegator API to accept a NameTree
Problem

The Delegator API accepts a `DelegatorTree[Name.Path]`, but this
isn't really appropriate: nothing has been delegated yet! This is
awkwar when constructing an initial DelelgatorTree for requests.

Solution

To fix this, the Delegator API is modified to accept a
`NameTree[Name.Path]`, as is this is actually the input we want.

Note: NameTree[Path] is accepted so that users may evaluate
expressions like `/s/foo | /s/bar`.