Change the internal Delegator API to accept a NameTree (#951)
Problem

The Delegator API accepts a `DelegatorTree[Name.Path]`, but this isn't really
appropriate: nothing has been delegated yet! This is awkward when constructing
an initial DelelgatorTree for requests.

Solution

To fix this, the Delegator API is modified to accept a `NameTree[Name.Path]`,
as is this is actually the input we want.