Introduce the `roundRobin` load balancer config type (#989)
Finagle's round-robin load balancer was not exposed as a configuration
option. Now it is is exposed as the `roundRobin` load balancer type.
This load balancer may be configured with a `maxEffort`, that controls
the number of times the load balancer can retry if the previously
picked node was marked unavailable.