Collapse client and server tracing spans
Change linkerd's handling of span IDs to more closely match Finagle's.
When linkerd receives a request, it will no longer advance the request's
span ID if it is already set. Instead, it will attach server annotations
to the caller's client span, and only advance the span ID when preparing
to make a new client request.