[h2] Replace io.netty.handler.codec.http2.Http2FrameCodec
Problem

Netty's Http2FrameCodec is a nascent API, and hides details that we really want
to control in linkerd: specifically flow control and http2 settings.  In order
to enact this sort of configuration changes we need, we'll need to use the
older Http2Connection API; but all of our transport code uses the new
Http2StreamFrame types.

Solution

Copy netty's Http2FrameCodec as H2FrameCodec, which reimplements the netty
codec in Scala. Now that we control this layer within linkerd, we'll be able to
make the needed configuration changes.  Once we're clearer what we need out of
this, we can make the needed changes to the original netty codec.

Related: #1013