[h2] Acknowledge window updates aggressively (#1031)
Problem

By default, netty does not send a WINDOW_UPDATE until the available window is
at ≥50% capacity. This can lead to starvation (especially for gRPC).

Solution

Change the default window increment ratio from 0.5 to 0.95 so that
WINDOW_UPDATE frames are sent much more aggressively.

This branch has been tested against the ver/grpc-interop-server branch (#1024,
to be merged soon after this), confirming that this fixes #1013.

In a follow-up, these parameters will be made configurable.