[grpc] Implement gRPC interop tests (#1024)
Various gRPC implementations, including grpc-go and grpc-java, provide default
implementations of the grpc.testing.TestService grpc service, as well as a CLI
that drives these tests.

This change provides an implementation for the TestService as well as a client
side that implements a subset of the standard interop tests.

The client and server are used to run these tests in CI. Two test
implementations are provided: a LocalInteropTest that just uses the interfaces
directly without an HTTP/2 transport, and a NetworkedInteropTest that exercises
these features with the full grpc-runtime, finagle-h2, and netty
implementation.

Unimplemented tests include:

- auth tokens
- custom metadata
- stream deadlines

Note: assembly doesn't work as-is. client and server configurations should be
added to support this

Co-authored-by: stevej@buoyant.io