Fix handling of failure accrual default
If a client config section does not include a failureAccrual property, the
default failure accrual policy will be applied, overriding any failure accrual
policy set in previous matching client configs.

We refactor slightly to use .maybeWith so that the failureAccrual param will
only be added if the property is defined.  A default value is supplied as a
router config that can be overwritten.

A test case in ClientTest validates this.