Classify exceptions properly in the server stack
When a request results in an exception, there is no response object to carry
the "l5d-success-class" header which is used for server stats.  This causes
the server stack to use the default response classifier and classify the
resulting HTTP 503s as successes.

Change the default response classifier to fall back to the
ServerErrorsAsFailures classifer so that exceptions get correctly classified
as failures.

A new test case in HttpEndToEndTest validates this.