Discard responses when retrying
When an HTTP response with a chunked body is classified as a retryable failure
and is retried, the chunked body is never read and the request is discarded.
The connection is never released because the response has unread chunks.  This
results in a connection leak.

Introduce a ResponseDiscarder stack param which is invoked on responses when
the RetryFilter decides to retry and thus discard the response.  For HTTP we
supply a ResponseDiscarder that explicitly discards chunked bodies.

Tested manually against a node.js server that returns chuncked 500s and added
an e2e test.  Shout out to taer for providing the node.js server and fantayeneh
for providing the test case.