Ensure h2 streams are fully received before closing (#1245)
Problem

A race condition existed that would cause a stream to be closed by the
remote sender before the local receiver had fully received all frames in
an h2 stream. This would result in the final frame of an h2 response
never being sent back to the caller, causing the caller's connection to
hang indefinitely.

Solution

Make the receiver fully responsible for closing streams, to ensure that
they are fully received before closing.

Validation

I've deployed this branch to a test environment and it has successfully
served gRPC traffic for the past 48 hours, whereas without this fix the
connection would always hang within about 20 minutes.