Fix flakey UsageDataTelemeterEndToEndTest
A test in UsageDataTelemeterEndToEndTest will nondeterministically fail if the
randomly jittered start delay of the UsageDataTelemeter is longer than the
await timeout.  The await timeout in CI is 30 seconds and the start delay is
uniformly random up to 1 minute.  This causes the test to fail approximately
50% of the time.

Use Time.withCurrentTimeFrozen and TimeControl to simulate enough time passing
to get past the start delay.