Add path stack registry
There is little visibility into what path stacks have been created and how they
have been configured.

Add a path stack registry that exposes information about the modules in each
path stack and the params that configure them.  This information is exposed
in the global registry which can be viewed at /admin/registry.json on the
admin site.  Also improve the readability of the ResponseClassifier and
Trust.Config params.