Namespaced Kubernetes Namer
Problem:
Users have requested a version of the k8s namer that is restricted
to only route to services in the namespace where linkerd is currently
running.

Solution:
This version of the k8s namer reads the namespace name from the
downward api and uses that namespace instead of specifying the
namespace in the dtab. Fixes #1199