Expire metrics for idle services and clients
If a service or client is idle for long enough (10 minutes by default), the
ServiceFactory will be closed and removed from cache but the metrics for that
service or client will live on forever.

We add the ability to prune MetricsTrees and delete all metrics for a service
or client when that ServiceFactory is closed.  We also expose a config option
for setting the ttl for idle ServiceFactories.

Tested manually.  Unable to create an end-to-end test because IdlingFactory
uses Stopwatch to determine idle time which does not respect TimeControl.