Client stats should use service response classifier (#1266)
If a response classifier is configured on a service, that response classifier
is not used to report client level success and failure metrics.  This can lead
to inconsistent metric counts.

Modify the PerDstStatsFilter to use the response classifier from the service
(stored in a local context).  Also add io.l5d.{http, h2}.allSuccessful response
classifiers to aid with testing and for general usefulness.