Port numbers in k8s names will now have the service's port mapping applied. (#1166)
When a port number is specified in a k8s name such as /#/io.l5d.k8s/ns/80/srv
the resulting address list will simply use the given port on each endpoint,
ignoring any port mapping specified in the service definition.

The port and targetPort from the service definition is taken into account when
evaluating k8s names with port numbers.  If a port number is specified that has
a target port, the name will resolve using the target port.  If a port number is
specified that has no target port, the target port is assumed to be the same as
the service port.  If a port number is specified that does not exist in the
service definition, Neg is returned.  This should result in more compatible
behavior when linkerd is used as an ingress controller and port numbers are
specified in the ingress resource.

Note that the targetPort is not available from the EndpointPort resource in the
endpoints API. Therefore, we must establish a watch on the services API in
addition to the watch on the endpoints API.