Stop tracking all closed streams in h2 dispatchers (#1215) (#1292)
* Stop tracking all closed streams in h2 dispatchers (#1215)

Problem

Netty4DispatcherBase maintains a ConcurrentHashMap with all known stream
states, including closed streams. This map grows unbounded as the number
of closed streams increase over the lifetime of a connection, causing a
memory leak.

Solution

Remove closed streams from the map, and store the highest closed stream
id. Streams with an id <= to the max and not found in the map are
assumed to be closed.

Validation

Tested locally and verified that the memory leak does not happen.

* Add recursive addClosedId func for setting max id