Mutual auth
While TLS server auth allowed clients to verify the identity of the server
that they are talking to, there is no way for servers to verify the identity
of clients that are talking to them.

Add `requireClientAuth` property to the server TLS config to require that only
clients that provide valid certificates may issue requests to that server.
Add a `clientAuth` section to the client TLS config where client credentials
can be provided.

Manually tested and integration test added.