Update h2 headers API with method for fetching single header value (#1298)
* Update h2 headers API with method for fetching single header value

Problem

In a local CPU profile, Netty4Message.Headers.get showed up as a minor
hotspot (0.1%). The method returns all values for a header, but in many
of the places where we call it, we immediately follow the call with a
`headOption` or `lastOption` call.

Solution

Rename the existing method to `getAll`, and add a new `get` method that
returns a single option value.

Validate

With this change, neither `get` or `getAll` show up as hotspots, but
this could also just be a result of splitting 1 method into two. That said,
I think the new API is cleaner and would like to keep it.

* Remove partial function from get method

* Switch getAll back to collect